requirements.txt
streamlit>=1.35
google-generativeai>=0.7
reportlab>=4.2
Pillow>=10.0
requests>=2.31
